LPS-167066 Move ObjectActionsFDSPropsTransformer to new directory
