+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/object_definitions/object_definition/actions.jsp
@@ -35,7 +35,7 @@ renderResponse.setTitle(objectDefinition.getLabel(locale, true));
 	fdsActionDropdownItems="<%= objectDefinitionsActionsDisplayContext.getFDSActionDropdownItems() %>"
 	formName="fm"
 	id="<%= ObjectDefinitionsFDSNames.OBJECT_ACTIONS %>"
-	propsTransformer="js/ObjectDefinitionsActionsFDSPropsTransformer"
+	propsTransformer="js/components/FDSPropsTransformer/ObjectActionsFDSPropsTransformer"
 	style="fluid"
 />
